Added AttributeSet & Refactored directories
Added the component "AttributeSet" that handles values like Health and other resources.

Updated directory hierarchy
