+++ Source/Project_D/Character/PlayableCharacter.cpp
@@ -8,6 +8,7 @@ APlayableCharacter::APlayableCharacter()
 {
 	PrimaryActorTick.bCanEverTick = true;
    AbilitySystemComponent = CreateDefaultSubobject<UAbilitySystem>(TEXT("AbilitySystemComponent"));
+   Attributes = CreateDefaultSubobject<UAttributeSet>(TEXT("Attributes"));
 }
 
 void APlayableCharacter::SetupPlayerInputComponent(UInputComponent* PlayerInputComponent)
