+++ Source/Project_D/Components/AbilitySystem/AbilitySystem.h
@@ -4,7 +4,7 @@
 
 #include "CoreMinimal.h"
 #include "Components/ActorComponent.h"
-#include "../Abilities/Ability.h"
+#include "../../Abilities/Ability.h"
 #include "AbilitySystem.generated.h"
 
 class ACharacterBase;
