+++ Source/Project_D/Interfaces/EffectHandler/EffectHandler.h
@@ -4,7 +4,7 @@
 
 #include "CoreMinimal.h"
 #include "UObject/Interface.h"
-#include "../GameplayEffect/GameplayEffect.h"
+#include "../../GameplayEffect/GameplayEffect.h"
 #include "EffectHandler.generated.h"
 
 // This class does not need to be modified.
